♻️ フェーズ2: CRUD操作の独立したフック化
- useBookDataフックを作成してデータ管理機能を独立化
- useBookMutationsフックを作成してCRUD操作を独立化
- books-providerからデータ取得とミューテーション機能を分離
- 型安全性を保ちながら責任を明確に分割
